Move edition validation to wks create/reset cmds
We actually want to do that logic everytime a
workspace is created or reset, not only when it
is done throught the `vtex use` command.

As the `use` command also already internally calls
the create and/or reset commands, we don't even need
any actual change to its logic. Still kept some of
the changes I ended up doing there as I believe
they are still improvements.
